pyproject.toml
[tool.poetry]
authors = ["tynguyen <tynguyen.tech@gmail.com>"]
description = ""
exclude = [
  "results",
  "cascade_fer/data",
  "cascade_fer/results",
  "cascade_fer/logs",
]
name = "cascade_fer"
readme = "README.md"
version = "0.1.1"

[tool.poetry.dependencies]
beautifulsoup4 = "^4.10.0"
dlib = "^19.22.1"
imutils = "^0.5.4"
keyring = "23.2.1"
"lazr.restfulclient" = ">=0.9.19"
matplotlib = "^3.4.3"
opencv-python = ">=4.0.0"
python = "^3.7"
scipy = ">=1.6.0"
tensorflow-gpu = "2.5"
urllib3 = ">=1.26.0"
idna = "^3.3"

[tool.poetry.dev-dependencies]
pytest = "~5"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.pytest.ini_options]
addopts = "-ra -q"
minversion = "0.6.0"
testpaths = ["tests"]
